requirements.txt
fastapi==0.103.0
uvicorn==0.22.0
python-socketio==5.9.0
caninos-sdk==0.5.0
gunicorn==21.2.0
